Replace V2 forms with updated versions
Removed LoginFormV2 and RegisterFormV2 along with their tests and styles. Updated imports to use the improved LoginForm and RegisterForm components. Simplified form handling by consolidating to a unified BaseAuthForm structure.
