+++ frontend-next-migration/src/features/AuthByUsername/index.ts
@@ -1,4 +1,2 @@
-// export { LoginForm } from './ui/LoginForm/LoginForm';
-export { LoginForm } from './ui/LoginFormV2/LoginFormV2';
-// export { RegisterForm } from './ui/RegisterForm/RegisterForm';
-export { RegisterForm } from './ui/RegisterFormV2/RegisterFormV2';
+export { LoginForm } from './ui/LoginForm/LoginForm';
+export { RegisterForm } from './ui/RegisterForm/RegisterForm';
