--- frontend-next-migration/src/features/AuthByUsername/ui/LoginForm/LoginForm.module.scss
@@ -1,5 +1,0 @@
-.registerLink{
-  text-align: center;
-  margin-top: 10px !important;
-  @include link-scale-effect;
-}
