+++ frontend-next-migration/src/features/AuthByUsername/ui/LoginForm/LoginForm.tsx
@@ -1,47 +1,52 @@
 'use client';
-import { CustomForm } from '@/shared/ui/CustomForm';
-import { AppLink, AppLinkTheme } from '@/shared/ui/AppLink/AppLink';
 import { useClientTranslation } from '@/shared/i18n';
 import { useLoginForm } from '../../model/useLoginForm';
-import cls from './LoginForm.module.scss';
+import { AppLink, AppLinkTheme } from '@/shared/ui/AppLink/AppLink';
+import { BaseAuthForm } from '@/entities/Auth';
 
 type Props = {
-    toForgottenPwPage: string;
+    toForgottenPwPage?: string;
     toRegisterPage: string;
     onSuccessLogin: () => void;
 };
 
-export const LoginForm = ({ toRegisterPage, onSuccessLogin }: Props) => {
-    const { t } = useClientTranslation('auth');
+export const LoginForm = (props: Props) => {
+    const { toRegisterPage, onSuccessLogin } = props;
 
+    const { t } = useClientTranslation('auth');
     const { register, handleSubmit, onFormSubmit, errors } = useLoginForm({ onSuccessLogin });
 
     return (
-        <CustomForm
-            className={cls.Form}
+        <BaseAuthForm
+            header={t('log_in')}
+            fields={
+                <>
+                    <BaseAuthForm.InputField
+                        key={'username'}
+                        error={errors?.username?.message && t(`${errors.username.message}`)}
+                        label={t('username')}
+                        inputProps={{ ...register('username'), required: true }}
+                    />
+                    <BaseAuthForm.InputField
+                        key={'password'}
+                        error={errors?.password?.message && t(`${errors.password.message}`)}
+                        label={t('password')}
+                        inputProps={{ ...register('password'), type: 'password', required: true }}
+                    />
+                </>
+            }
+            actions={
+                <>
+                    <BaseAuthForm.SubmitButton>{t('send')}</BaseAuthForm.SubmitButton>
+                    <AppLink
+                        theme={AppLinkTheme.PRIMARY}
+                        to={toRegisterPage}
+                    >
+                        {t('text_to_register')}
+                    </AppLink>
+                </>
+            }
             onSubmit={handleSubmit(onFormSubmit)}
-        >
-            <CustomForm.Header>{t('log_in')}</CustomForm.Header>
-            <CustomForm.InputField
-                key={'username'}
-                error={errors?.username?.message && t(`${errors?.username?.message}`)}
-                label={t('username')}
-                inputProps={{ ...register('username'), required: true }}
-            />
-            <CustomForm.InputField
-                key={'password'}
-                error={errors?.password?.message && t(`${errors?.password?.message}`)}
-                label={t('password')}
-                inputProps={{ ...register('password'), type: 'password', required: true }}
-            />
-            <CustomForm.Button type="submit">{t('send')}</CustomForm.Button>
-            <AppLink
-                theme={AppLinkTheme.PRIMARY}
-                to={toRegisterPage}
-                className={cls.registerLink}
-            >
-                {t('text_to_register')}
-            </AppLink>
-        </CustomForm>
+        />
     );
 };
