--- frontend-next-migration/src/features/AuthByUsername/ui/RegisterForm/RegisterForm.module.scss
@@ -1,5 +1,0 @@
-.loginLink{
-  text-align: center;
-  margin-top: 10px !important;
-  @include link-scale-effect
-}
